backend/transaction_microservice/requirements.txt
flask
requests
firebase-admin
stripe
python-dotenv
